split identifier from complete.ts
