+++ packages/server/src/complete/utils.ts
@@ -1,0 +1,6 @@
+export function makeTableAlias(tableName: string): string {
+  if (tableName.length > 3) {
+    return tableName.substring(0, 3)
+  }
+  return tableName
+}
